third_party/WebKit/LayoutTests/svg/filters/feComponentTransfer-style-crash-expected.txt

ALERT: SVG_INVALID_VALUE_ERR: An invalid value was passed to an operation or assigned to an attribute.
PASS
